fix: convert paths to relative for vitePreprocess sourcemap sources
